move exported functions to the top-level src folder
