+++ src/compose.js
